+++ e2e/downloads/fileSupported.spec.ts
@@ -14,20 +14,31 @@ test('supported file types are handled correctly', async ({ page }) => {
     '"href": "https://www.iana.org/assignments/media-types/application/geopackage+sqlite3"'
   )
 
+  // GeoPackages (Zipped)
+  //
+
+  await page.click('text=Add Download')
+  await page.click('text=Local File')
+  await page.setInputFiles('input#download-7-file', './sample-data/gpkg_zip/sample.gpkg.zip')
+  const download2 = await page.textContent('#download-7-output pre')
+  expect(download2).toContain(
+    '"href": "https://metadata-resources.data.bas.ac.uk/media-types/application/geopackage+sqlite3+zip"'
+  )
+
   // JPEGs
   //
 
   await page.click('text=Add Download')
   await page.click('text=Local File')
   await page.setInputFiles('input#download-2-file', './sample-data/jpeg/sample.jpg')
-  const download2 = await page.textContent('#download-2-output pre')
-  expect(download2).toContain('"href": "https://jpeg.org/jpeg/"')
+  const download3 = await page.textContent('#download-2-output pre')
+  expect(download3).toContain('"href": "https://jpeg.org/jpeg/"')
 
   await page.click('text=Add Download')
   await page.click('text=Local File')
   await page.setInputFiles('input#download-3-file', './sample-data/jpeg/sample.jpeg')
-  const download3 = await page.textContent('#download-3-output pre')
-  expect(download3).toContain('"href": "https://jpeg.org/jpeg/"')
+  const download4 = await page.textContent('#download-3-output pre')
+  expect(download4).toContain('"href": "https://jpeg.org/jpeg/"')
 
   // PNGs
   //
@@ -35,8 +46,8 @@ test('supported file types are handled correctly', async ({ page }) => {
   await page.click('text=Add Download')
   await page.click('text=Local File')
   await page.setInputFiles('input#download-4-file', './sample-data/png/sample.png')
-  const download4 = await page.textContent('#download-4-output pre')
-  expect(download4).toContain('"href": "https://www.iana.org/assignments/media-types/image/png"')
+  const download5 = await page.textContent('#download-4-output pre')
+  expect(download5).toContain('"href": "https://www.iana.org/assignments/media-types/image/png"')
 
   // PDFs
   //
@@ -44,9 +55,20 @@ test('supported file types are handled correctly', async ({ page }) => {
   await page.click('text=Add Download')
   await page.click('text=Local File')
   await page.setInputFiles('input#download-5-file', './sample-data/pdf/sample.pdf')
-  const download5 = await page.textContent('#download-5-output pre')
-  expect(download5).toContain(
+  const download6 = await page.textContent('#download-5-output pre')
+  expect(download6).toContain(
     '"href": "https://www.iana.org/assignments/media-types/application/pdf"'
+  )
+
+  // PDFs (geo-referenced)
+  //
+
+  await page.click('text=Add Download')
+  await page.click('text=Local File')
+  await page.setInputFiles('input#download-5-file', './sample-data/pdf/sample.pdf')
+  const download7 = await page.textContent('#download-5-output pre')
+  expect(download7).toContain(
+    '"href": "https://www.iana.org/assignments/media-types/application/pdf+geo https://metadata-resources.data.bas.ac.uk/media-types/application/pdf+geo"| '
   )
 
   // Shapefiles (Zipped)
@@ -55,19 +77,8 @@ test('supported file types are handled correctly', async ({ page }) => {
   await page.click('text=Add Download')
   await page.click('text=Local File')
   await page.setInputFiles('input#download-6-file', './sample-data/shp_zip/sample.shp.zip')
-  const download6 = await page.textContent('#download-6-output pre')
-  expect(download6).toContain(
-    '"href": "https://metadata-standards.data.bas.ac.uk/media-types/application/shapefile+zip"'
-  )
-
-  // GeoPackages (Zipped)
-  //
-
-  await page.click('text=Add Download')
-  await page.click('text=Local File')
-  await page.setInputFiles('input#download-7-file', './sample-data/gpkg_zip/sample.gpkg.zip')
-  const download7 = await page.textContent('#download-7-output pre')
-  expect(download7).toContain(
-    '"href": "https://metadata-standards.data.bas.ac.uk/media-types/application/geopackage+sqlite3+zip"'
+  const download8 = await page.textContent('#download-6-output pre')
+  expect(download8).toContain(
+    '"href": "https://metadata-resources.data.bas.ac.uk/media-types/application/shapefile+zip"'
   )
 })